add viewer routes for VFEs/scenes
