+++ src/App.tsx
@@ -5,7 +5,8 @@ import CreateVFEForm from "./CreateVFE.tsx";
 import { VFE } from "./DataStructures.ts";
 import LandingPage from "./LandingPage.tsx";
 import PhotosphereEditor from "./PhotosphereEditor.tsx";
-import App from "./Prototype.tsx";
+import PhotosphereLoader from "./PhotosphereLoader.tsx";
+import Prototype from "./Prototype.tsx";
 
 // Main component acts as a main entry point for the application
 // Should decide what we are doing, going to LandingPage/Rendering VFE
@@ -55,7 +56,10 @@ function AppRoot() {
           />
         }
       />
-      <Route path="/viewer" element={<App />} />
+      <Route path="/viewer" element={<Prototype />} />
+      <Route path="/viewer/:vfeID" element={<PhotosphereLoader />}>
+        <Route path=":sceneID" />
+      </Route>
       <Route
         path="/create"
         element={<CreateVFEForm onCreateVFE={loadCreatedVFE} />}
